Reorganize C++ resources
